Refactor job summary rendering for support status
Includes other minor cleanups
