+++ sources/test/jest/job-summary.test.ts
@@ -11,7 +11,7 @@ jest.unstable_mockModule('../../src/wrapper-validation/wrapper-checksums.json', 
     }
 }))
 
-const {jobMarker, renderSummaryTable} = await import('../../src/job-summary')
+const {jobMarker, renderSummaryTable, renderSupportStatusUsing} = await import('../../src/job-summary')
 
 async function renderWith(versions: string[], results: BuildResult[]): Promise<string> {
     releasedVersions = versions.map(version => ({version, checksum: ''}))
@@ -21,6 +21,30 @@ async function renderWith(versions: string[], results: BuildResult[]): Promise<s
 }
 
 const DOC = 'https://docs.gradle.org/current/userguide/feature_lifecycle.html#eol_support'
+const SECURITY_SUBSCRIPTION = 'https://gradle.org/security-subscription/?utm_source=github-action'
+
+// Mirrors the real release shape: latest 9.7.1, last 8.x minor is 8.14 with patches up to 8.14.5.
+const RELEASED = [
+    '9.7.1',
+    '9.7.0',
+    '9.6.1',
+    '9.6.0',
+    '9.5.1',
+    '9.5.0',
+    '9.4.1',
+    '9.4.0',
+    '9.2.1',
+    '9.2.0',
+    '9.0.0',
+    '8.14.5',
+    '8.14',
+    '8.13',
+    '8.3',
+    '8.0.2',
+    '8.0',
+    '7.6.4',
+    '1.0'
+]
 
 const MATRIX_INPUT_ENV = 'INPUT_WORKFLOW-JOB-CONTEXT'
 
@@ -47,19 +71,20 @@ const failedHelpBuild: BuildResult = {
 
 const longArgsBuild: BuildResult = {
     ...successfulHelpBuild,
-    requestedTasks: 'check publishMyLongNamePluginPublicationToMavenCentral publishMyLongNamePluginPublicationToPluginPortal',
+    requestedTasks:
+        'check publishMyLongNamePluginPublicationToMavenCentral publishMyLongNamePluginPublicationToPluginPortal'
 }
 
 const scanPublishDisabledBuild: BuildResult = {
     ...successfulHelpBuild,
     buildScanUri: '',
-    buildScanFailed: false,
+    buildScanFailed: false
 }
 
 const scanPublishFailedBuild: BuildResult = {
     ...successfulHelpBuild,
     buildScanUri: '',
-    buildScanFailed: true,
+    buildScanFailed: true
 }
 
 describe('renderSummaryTable', () => {
@@ -83,7 +108,7 @@ describe('renderSummaryTable', () => {
                         <td><a href="https://scans.gradle.com/s/abc123" rel="nofollow" target="_blank"><img src="https://img.shields.io/badge/Build%20Scan%C2%AE-06A0CE?logo=Gradle" alt="Build Scan published" /></a></td>
                     </tr>
                 </table>
-            `);
+            `)
         })
         it('failed build', () => {
             const table = renderSummaryTable([failedHelpBuild])
@@ -104,7 +129,7 @@ describe('renderSummaryTable', () => {
                         <td><a href="https://scans.gradle.com/s/abc123" rel="nofollow" target="_blank"><img src="https://img.shields.io/badge/Build%20Scan%C2%AE-06A0CE?logo=Gradle" alt="Build Scan published" /></a></td>
                     </tr>
                 </table>
-            `);
+            `)
         })
         describe('when build scan', () => {
             it('publishing disabled', () => {
@@ -126,7 +151,7 @@ describe('renderSummaryTable', () => {
                             <td><a href="https://scans.gradle.com" rel="nofollow" target="_blank"><img src="https://img.shields.io/badge/Not%20published-lightgrey" alt="Build Scan not published" /></a></td>
                         </tr>
                     </table>
-                `);
+                `)
             })
             it('publishing failed', () => {
                 const table = renderSummaryTable([scanPublishFailedBuild])
@@ -147,7 +172,7 @@ describe('renderSummaryTable', () => {
                             <td><a href="https://docs.gradle.com/develocity/gradle-plugin/#troubleshooting" rel="nofollow" target="_blank"><img src="https://img.shields.io/badge/Publish%20failed-orange" alt="Build Scan publish failed" /></a></td>
                         </tr>
                     </table>
-                `);
+                `)
             })
         })
         it('multiple builds', () => {
@@ -176,7 +201,7 @@ describe('renderSummaryTable', () => {
                         <td><a href="https://scans.gradle.com/s/abc123" rel="nofollow" target="_blank"><img src="https://img.shields.io/badge/Build%20Scan%C2%AE-06A0CE?logo=Gradle" alt="Build Scan published" /></a></td>
                     </tr>
                 </table>
-            `);
+            `)
         })
         it('truncating long requested tasks', () => {
             const table = renderSummaryTable([longArgsBuild])
@@ -197,7 +222,7 @@ describe('renderSummaryTable', () => {
                         <td><a href="https://scans.gradle.com/s/abc123" rel="nofollow" target="_blank"><img src="https://img.shields.io/badge/Build%20Scan%C2%AE-06A0CE?logo=Gradle" alt="Build Scan published" /></a></td>
                     </tr>
                 </table>
-            `);
+            `)
         })
     })
 })
@@ -226,7 +251,7 @@ describe('Gradle version support status', () => {
             <details>
                 <summary>:warning: Gradle 8.0 is end-of-life</summary>
                 <p>The 8.x release line receives no new fixes of any kind. Update to the latest Gradle version.</p>
-                <p>Options for staying secure on an end-of-life version: <a href="https://gradle.org/security-subscription/?utm_source=github-action">Gradle Security Subscription</a></p>
+                <p>Options for staying secure on an end-of-life version: <a href="${SECURITY_SUBSCRIPTION}">Gradle Security Subscription</a></p>
             </details>
         `)
     })
@@ -243,6 +268,57 @@ describe('Gradle version support status', () => {
         expect(table).toContain(`<td align='center'>8.0</td>`)
         expect(table).not.toContain(':information_source:')
         expect(table).not.toContain('Consider upgrading')
+    })
+})
+
+describe('renderSupportStatus', () => {
+    const LEGEND = `<p>:information_source: Consider upgrading — See <a href="${DOC}">Gradle release lifecycle</a></p>`
+
+    it('folds an end-of-life version under a warning sign, linking the security subscription', () => {
+        const rendered = renderSupportStatusUsing(['7.6.4'], RELEASED)
+
+        expect(rendered).toContain('<summary>:warning: Gradle 7.6.4 is end-of-life</summary>')
+        expect(rendered).toContain(
+            'The 7.x release line receives no new fixes of any kind. Update to the latest Gradle version.'
+        )
+        expect(rendered).toContain(`<a href="${SECURITY_SUBSCRIPTION}">Gradle Security Subscription</a>`)
+        expect(rendered).not.toContain(LEGEND)
+    })
+
+    it('names no version in the fold beyond the one that is end-of-life', () => {
+        expect(renderSupportStatusUsing(['7.6.4'], RELEASED)).not.toContain('9.7.1')
+    })
+
+    it.each(['8.0', '8.14.3', '8.14.5', '9.2.1'])('adds only the legend for %s', version => {
+        expect(renderSupportStatusUsing([version], RELEASED).trim()).toBe(LEGEND)
+    })
+
+    it('adds the legend exactly once however many versions are flagged', () => {
+        const rendered = renderSupportStatusUsing(['9.2.1', '9.4.1', '8.14.5', '8.14.3', '8.3', '8.0'], RELEASED)
+
+        expect(rendered.trim()).toBe(LEGEND)
+    })
+
+    it('emits one fold per end-of-life version plus the single legend', () => {
+        const rendered = renderSupportStatusUsing(['9.6.1', '9.2.1', '8.14.5', '8.0', '7.6.4', '1.0'], RELEASED)
+
+        expect(rendered.match(/<details>/g)).toHaveLength(2)
+        expect(rendered.match(/Consider upgrading/g)).toHaveLength(1)
+    })
+
+    it('names no version outside the fold', () => {
+        const rendered = renderSupportStatusUsing(['8.0', '9.2.1'], RELEASED)
+
+        expect(rendered).not.toContain('8.0')
+        expect(rendered).not.toContain('9.2.1')
+    })
+
+    it('renders nothing when every version is inside the grace band', () => {
+        expect(renderSupportStatusUsing(['9.5.1', '9.6.1', '9.7.1'], RELEASED)).toBe('')
+    })
+
+    it('renders nothing when no Gradle build ran', () => {
+        expect(renderSupportStatusUsing([], RELEASED)).toBe('')
     })
 })
